[TDF] Move ColumnName2ColumnTypeName and CreateAction to TDataFrameInterface.hxx
